+++ src/krak/intern/client.clj
@@ -8,6 +8,8 @@
 
 (def KRAKEN-API "https://api.kraken.com")
 
+(defn comma-join [coll] (if (empty? coll) nil (clojure.string/join "," coll)))
+
 ;; API-Key = API key
 ;; API-Sign = Message signature using HMAC-SHA512 of
 ;;   (URI path + SHA256(nonce + POST data)) and base64 decoded secret API key
@@ -17,12 +19,7 @@
         path-digest (byte-array (concat (.getBytes path) nd-digest))
         hmac-bytes (sha512-hmac-bytes path-digest secret-bytes)
         signature (String. (b64/encode hmac-bytes))]
-    {"API-KEY" api-key
-     "API-Sign" signature}))
-
-(defn cj [coll]
-  (if (empty? coll) nil
-      (clojure.string/join "," coll)))
+    {"API-KEY" api-key "API-Sign" signature}))
 
 (defn- prepare-opt [[key {:keys [optional prepare-fn]}] data]
   (if (and (not optional) (nil? data))
@@ -39,28 +36,26 @@
         ik (set (keys inputs))]
     (if-not (set/subset? ok ik)
      (throw (Exception. (str "unknown arguments: " (set/difference ok ik))))))
-  (let [pairs (remove (comp nil? second)
-                      (map #(prepare-opt %
-                                         ((first %) opts))
-                           inputs))]
-    (into {} pairs)))
+  (into {} (remove (comp nil? second)
+                   (map #(prepare-opt % ((first %) opts)) inputs))))
 
-(defn- endpoint-fn [path inputs signature]
-  (if signature
-    (fn [ctx & {:as opts}]
-      (let [{:keys [api-key secret-key nonce-offset]} @ctx
-            nonce (+ nonce-offset (System/currentTimeMillis))
-            data (#'http/query-string
-                  (assoc (prepare-opts inputs opts) :nonce nonce))
-            headers (sign-request api-key secret-key path data nonce)]
-        (swap! ctx update :rate-counter + 2)
-        (let [response @(http/post (str KRAKEN-API path)
-                                   {:body data :headers headers})]
-          (assoc response :data (json/read-str (:body response))))))
-    (fn [& {:as opts}]
-      (let [data (#'http/query-string (prepare-opts inputs opts))
-            response @(http/post (str KRAKEN-API path) {:body data})]
-        (assoc response :data (json/read-str (:body response)))))))
+(defn- send-req [path data headers]
+  (let [response @(http/post (str KRAKEN-API path)
+                             {:body data :headers headers})
+        resp-data (json/read-str (:body response))]
+    (merge response {:data resp-data
+                     :krak/result (resp-data "result")
+                     :krak/error (resp-data "error")})))
 
-(defn register-endpoint [{:keys [fn-name path inputs signature]}]
-  (intern *ns* fn-name (endpoint-fn path inputs signature)))
+(defn private-req [path inputs]
+  (fn [ctx & {:as opts}]
+    (let [{:keys [api-key secret-key nonce-offset]} @ctx
+          nonce (+ nonce-offset (System/currentTimeMillis))
+          data (#'http/query-string
+                (assoc (prepare-opts inputs opts) :nonce nonce))
+          headers (sign-request api-key secret-key path data nonce)]
+      (send-req path data headers))))
+
+(defn public-req [path inputs]
+  (fn [& {:as opts}]
+    (send-req path (#'http/query-string (prepare-opts inputs opts)) {})))