+++ CHANGELOG.md
@@ -7,9 +7,17 @@ All notable changes to this project will be documented in this file. This change
 ### Removed
 ### Fixed
 
+## [0.1.1] - 2017-09-20
+### Added
+- :krak/result and :krak/error are now included in response map
+### Changed
+- Define endpoints explicitly
+- Fixed license in project.clj
+
 ## [0.1.0] - 2017-09-19
 ### Added
 - Initial API implementation
 
-[Unreleased]: https://github.com/keymone/krak/compare/0.1.0...HEAD
-[0.1.0]: https://github.com/keymone/krak/compare/00000000...0.1.0
+[Unreleased]: https://github.com/keymone/krak/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/keymone/krak/compare/0.1.0...0.1.1
+[0.1.0]: https://github.com/keymone/krak/commit/ab829373